Successfully moving camera_axis, not getting any pos data yet

+++ cob_camera_axis/common/include/cob_camera_axis/ElmoCtrl.h
@@ -179,7 +179,7 @@ private:
 	//CANPeakSysUSB * m_CanCtrl;
 	CanESD * m_CanCtrl;
 
-	float m_MaxVel;
+	double m_MaxVel;
 	int m_HomingDir;
 
 	double  m_UpperLimit;
